requirements.txt
keras==3.1.1
numpy==1.26.4
Pillow==10.3.0
streamlit==1.32.2
